pyproject.toml
[project]
name = "pilotpm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.135.2",
    "google-genai>=1.69.0",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langgraph>=1.1.3",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "structlog>=25.5.0",
    "twilio>=9.10.4",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]
